Tests/LibWeb/Text/expected/HTML/label-click-event-properties.txt

target: input
clientX preserved from label event: true
clientY preserved from label event: true
offsetX relative to input: true
offsetY relative to input: true
